pyproject.toml
[tool.poetry]
name = "aerospace-chatbot"
version = "0.1.0"
description = ""
authors = ["Dan Mueller <dsm@danmueller.pro>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiobotocore = "2.8.0"
aiohttp = "3.9.1"
aioitertools = "0.11.0"
aiosignal = "1.3.1"
altair = "5.2.0"
anyio = "3.7.1"
appnope = "0.1.3"
asttokens = "2.4.1"
attrs = "23.1.0"
blinker = "1.7.0"
botocore = "1.33.1"
cachetools = "5.3.2"
canopy = "8.42"
canopy-sdk = "0.1.3"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
click = "8.1.7"
comm = "0.2.0"
dataclasses-json = "0.6.3"
debugpy = "1.8.0"
decorator = "5.1.1"
dnspython = "2.4.2"
executing = "2.0.1"
fastapi = "0.92.0"
frozenlist = "1.4.0"
fsspec = "2023.12.1"
gcsfs = "2023.12.1"
gitdb = "4.0.11"
gitpython = "3.1.40"
google-api-core = "2.15.0"
google-auth = "2.25.2"
google-auth-oauthlib = "1.1.0"
google-cloud-core = "2.4.1"
google-cloud-storage = "2.13.0"
google-crc32c = "1.5.0"
google-resumable-media = "2.6.0"
googleapis-common-protos = "1.62.0"
gunicorn = "21.2.0"
h11 = "0.14.0"
idna = "3.6"
importlib-metadata = "6.11.0"
ipykernel = "6.27.1"
ipython = "8.18.1"
jedi = "0.19.1"
jinja2 = "3.1.2"
jmespath = "1.0.1"
joblib = "1.3.2"
jsonpatch = "1.33"
jsonpointer = "2.4"
jsonschema = "4.20.0"
jsonschema-specifications = "2023.11.2"
jupyter-client = "8.6.0"
jupyter-core = "5.5.0"
langchain = "0.0.348"
langchain-core = "0.0.12"
langchainhub = "0.1.14"
langsmith = "0.0.69"
loguru = "0.7.2"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
marshmallow = "3.20.1"
matplotlib-inline = "0.1.6"
mdurl = "0.1.2"
mmh3 = "3.1.0"
multidict = "6.0.4"
munch = "4.0.0"
mypy-extensions = "1.0.0"
nest-asyncio = "1.5.8"
nltk = "3.8.1"
oauthlib = "3.2.2"
openai = "0.27.5"
packaging = "23.2"
pandas-stubs = "2.0.3.230814"
parso = "0.8.3"
pexpect = "4.9.0"
pillow = "10.1.0"
pinecone-client = "2.2.4"
pinecone-datasets = "0.6.2"
pinecone-text = "0.6.1"
platformdirs = "4.1.0"
prompt-toolkit = "3.0.41"
protobuf = "4.25.1"
psutil = "5.9.6"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pyarrow = "11.0.0"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pydantic = "1.10.13"
pydeck = "0.8.0"
pygments = "2.17.2"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
pyzmq = "25.1.2"
referencing = "0.32.0"
regex = "2023.10.3"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rich = "13.7.0"
rpds-py = "0.13.2"
rsa = "4.9"
s3fs = "2023.12.1"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.0"
sqlalchemy = "2.0.23"
sse-starlette = "1.8.2"
stack-data = "0.6.3"
starlette = "0.25.0"
streamlit = "1.29.0"
tenacity = "8.2.3"
tiktoken = "0.3.3"
toml = "0.10.2"
toolz = "0.12.0"
tornado = "6.4"
tqdm = "4.66.1"
traitlets = "5.14.0"
types-jsonschema = "4.20.0.0"
types-pytz = "2023.3.1.1"
types-pyyaml = "6.0.12.12"
types-requests = "2.31.0.10"
types-tqdm = "4.66.0.5"
typing-inspect = "0.9.0"
typing-extensions = "4.8.0"
tzdata = "2023.3"
tzlocal = "5.2"
urllib3 = "2.0.7"
uvicorn = "0.20.0"
validators = "0.22.0"
wcwidth = "0.2.12"
wget = "3.2"
wrapt = "1.16.0"
yarl = "1.9.4"
zipp = "3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
